requirements.txt
django==3.1.4
djangorestframework==3.12.2
Faker==5.3.0

